fixed MainLayout.tsx
